Refactor app router structure to separate docs and app
