+++ package.json
@@ -9,6 +9,8 @@
         "lint": "next lint"
     },
     "dependencies": {
+        "@mantine/core": "^7.15.1",
+        "@mantine/hooks": "^7.15.1",
         "@types/mdx": "^2.0.13",
         "@vercel/postgres": "^0.9.0",
         "fumadocs-core": "^14.6.1",
@@ -27,7 +29,9 @@
         "@types/ws": "^8.5.13",
         "eslint": "^9",
         "eslint-config-next": "15.1.0",
-        "postcss": "^8",
+        "postcss": "^8.4.49",
+        "postcss-preset-mantine": "^1.17.0",
+        "postcss-simple-vars": "^7.0.1",
         "tailwindcss": "^3.4.1",
         "typescript": "^5"
     }
