+++ src/app/(docs)/docs/[[...slug]]/page.tsx
