+++ src/app/(main)/dashboard/layout.tsx
