+++ src/app/(main)/dashboard/page.tsx
@@ -1,0 +1,10 @@
+import WideLogo from "@/components/header/WideLogo";
+
+// TODO: Add this component
+export default function Page() {
+    return (
+        <>
+            <WideLogo />
+        </>
+    );
+}
