+++ src/app/(main)/layout.tsx
@@ -1,0 +1,30 @@
+import type { Metadata } from "next";
+import "@mantine/core/styles.css";
+
+import {
+    ColorSchemeScript,
+    MantineProvider,
+    mantineHtmlProps,
+} from "@mantine/core";
+
+export const metadata: Metadata = {
+    title: "Create Next App",
+    description: "Generated by create next app",
+};
+
+export default function RootLayout({
+    children,
+}: Readonly<{
+    children: React.ReactNode;
+}>) {
+    return (
+        <html lang="en">
+            <head>
+                <ColorSchemeScript />
+            </head>
+            <body>
+                <MantineProvider>{children}</MantineProvider>
+            </body>
+        </html>
+    );
+}
