--- src/app/programs/page.tsx
@@ -1,4 +1,0 @@
-// TODO: Add this component
-export default function Page() {
-    return <></>;
-}
