+++ tailwind.config.js
@@ -3,15 +3,11 @@ import { createPreset } from "fumadocs-ui/tailwind-plugin";
 /** @type {import('tailwindcss').Config} */
 export default {
     content: [
-        "./components/**/*.{ts,tsx}",
         "./app/**/*.{ts,tsx}",
-        "./content/**/*.{md,mdx}",
-        "./node_modules/fumadocs-ui/dist/**/*.js",
-        "./src/pages/**/*.{js,ts,jsx,tsx,mdx}",
         "./src/components/**/*.{js,ts,jsx,tsx,mdx}",
         "./src/app/**/*.{js,ts,jsx,tsx,mdx}",
-        "./node_modules/fumadocs-ui/dist/**/*.js",
 
+        "./node_modules/fumadocs-ui/dist/**/*.js",
         "./docs/**/*.mdx",
         "./mdx-components.tsx",
     ],
